requirements.txt
antlr4-python3-runtime==4.8
certifi==2021.10.8
charset-normalizer==2.0.12
hydra-core==1.1.2
idna==3.3
omegaconf==2.1.2
PyYAML==6.0
requests==2.27.1
urllib3==1.26.9
